docs/sdks/s3/go: fixup the examples and ensure they all compile
Closes TIG-3469

This moves the example files into a separate folder so that we can
use editor tooling to ensure that they not only compile, but we have
tab-completion.

This will be done to the other languages when I continue plowing
through the examples.

All examples are known to work and have been tested locally.